+++ parapet-core/src/main/scala/io/parapet/core/processes/SystemProcess.scala
@@ -2,7 +2,7 @@ package io.parapet.core.processes
 
 import io.parapet.{Event, Process, ProcessRef}
 import io.parapet.ProcessRef.{DeadLetterRef, SystemRef}
-import io.parapet.core.Events.{DeadLetter, Failure}
+import io.parapet.Event.{DeadLetter, Failure}
 
 /** Built-in singleton process pinned to [[io.parapet.ProcessRef.SystemRef]] that handles runtime-level failures.
   *
