+++ parapet-net/src/main/scala/io/parapet/net/processes/ClientProcess.scala
@@ -1,7 +1,7 @@
 package io.parapet.net.processes
 
 import com.typesafe.scalalogging.Logger
-import io.parapet.core.Events.Stop
+import io.parapet.Event.Stop
 import io.parapet.{Event, Process}
 import io.parapet.net.transport.{Message, TransportError}
 import io.parapet.net.transport.ClientTransport
