+++ parapet-net/src/main/scala/io/parapet/net/processes/ServerProcess.scala
@@ -1,7 +1,7 @@
 package io.parapet.net.processes
 
 import com.typesafe.scalalogging.Logger
-import io.parapet.core.Events.{Start, Stop}
+import io.parapet.Event.{Start, Stop}
 import io.parapet.{Event, Process}
 import io.parapet.net.transport.{Message, ReceiveResult, RoutedMessage, RoutingId, TransportError}
 import io.parapet.net.transport.ServerTransport
