+++ parapet-pario/src/test/scala/io/parapet/tests/intg/pario/JournalRecordingIntgSpec.scala
@@ -1,6 +1,6 @@
 package io.parapet.tests.intg.pario
 
-import io.parapet.core.Events.Start
+import io.parapet.Event.Start
 import io.parapet.{ParConfig, Process}
 import io.parapet.effect.ParIO
 import io.parapet.effect.ParIO.given
