+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/DslSpec.scala
@@ -1,7 +1,7 @@
 package io.parapet.tests.intg
 
 import io.parapet.core.Dsl.DslF
-import io.parapet.core.Events.Start
+import io.parapet.Event.Start
 import io.parapet.tests.intg.DslSpec._
 import io.parapet.testutils.EventStore
 import io.parapet.{Event, Process, ProcessRef}
