+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/DynamicProcessCreationSpec.scala
@@ -1,6 +1,6 @@
 package io.parapet.tests.intg
 
-import io.parapet.core.Events.Start
+import io.parapet.Event.Start
 import io.parapet.tests.intg.DynamicProcessCreationSpec._
 import io.parapet.testutils.EventStore
 import io.parapet.{Event, Process, ProcessRef}
