+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/EventDeliverySpec.scala
@@ -1,6 +1,6 @@
 package io.parapet.tests.intg
 
-import io.parapet.core.Events._
+import io.parapet.Event._
 import io.parapet.core.exceptions.EventMatchException
 import io.parapet.core.processes.DeadLetterProcess
 import io.parapet.runtime.Envelope
