+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/SelfSendSpec.scala
@@ -1,7 +1,7 @@
 package io.parapet.tests.intg
 
 import io.parapet.{Event, Process}
-import io.parapet.core.Events.Start
+import io.parapet.Event.Start
 import io.parapet.tests.intg.SelfSendSpec._
 import io.parapet.testutils.EventStore
 import org.scalatest.flatspec.AnyFlatSpec
